Modified: split user authentication controller to controller & service according to controller handles http logic while service handles business logic
